CODES AND EPSILON VALUE FOR randInitializeWeights FUNCTION
This is a MATLAB file that contains the codes and the choice of the initial value of epsilon in the setting up of the randInitializeWeights function in order to achieve the process of symmetry breaking, by randomly initializing the parameters (weights) of the Neural Network. These weights govern the mapping of a function from one layer in the neural network to the next.