FEED FORWARD AND BACKPROPAGATION ALGORITHM CODES - NEURAL NETWORK CASE STUDY
This MATLAB file contains the codes (in the form of Matrices and vector operations) that I wrote to implement the Feed Forward and Backpropagation algorithms to train a Neural Network in the hand-written digit recognition problem. The codes in this file formed part of the course work undertaken for Stanford University's Machine Learning course, offered via Coursera.